requirements.txt
asgiref==3.9.2
attrs==25.3.0
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
